re #34 - Minor refactor
