+++ CasperVanBattum/Breakin/Assets/Scripts/Breakin/BasicBlock.cs
@@ -1,4 +1,4 @@
-﻿namespace Prototype1
+﻿namespace Breakin
 {
 	public class BasicBlock : Block
 	{
