+++ CasperVanBattum/Breakin/Assets/Scripts/Breakin/Pooling/BlockPool.cs
@@ -1,0 +1,10 @@
+namespace Breakin.Pooling
+{
+    public class BlockPooler
+    {
+        public BlockPooler(int capacity)
+        {
+            
+        }
+    }
+}
