+++ CasperVanBattum/Breakin/Assets/Scripts/Breakin/Pooling/IPoolable.cs
@@ -1,0 +1,7 @@
+namespace Breakin.Pooling
+{
+    public interface IPoolable
+    {
+        
+    }
+}
