+++ CasperVanBattum/Breakin/Assets/Scripts/Breakin/Ring.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Prototype1
+namespace Breakin
 {
 	public class Ring
 	{
