+++ CasperVanBattum/Breakin/Assets/Scripts/Breakin/ScoreManager.cs
@@ -1,30 +1,34 @@
 using System;
-using System.Runtime.CompilerServices;
-using UnityEngine;
 
-namespace Prototype1
+namespace Breakin
 {
     public interface IScoreManager
     {
         int Score { get; set; }
-        event Action OnScoreChanged;
+
+        /// <summary>
+        /// Invaked each time the Score property is changed
+        /// </summary>
+        event Action ScoreChanged;
     }
-    
+
     public class ScoreManager : IScoreManager
     {
         private static IScoreManager instance;
         public static IScoreManager Instance => instance ?? (instance = new ScoreManager());
 
-        public event Action OnScoreChanged;
+        public event Action ScoreChanged;
 
         private int score;
+
         public int Score
         {
             get => score;
             set
             {
                 score = value;
-                OnScoreChanged?.Invoke();
+                // Let subscribers know the score was changed
+                ScoreChanged?.Invoke();
             }
         }
     }
